requirements.txt
langchain-community
ollama
chromadb
streamlit
pypdf